Separate route file  into new files Co-authored-by: dudamelo82 <dudamelo82@gmail.com>
